+++ lib/deepl/exceptions/document_translation_error.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module DeepL
+  module Exceptions
+    class DocumentTranslationError < Error
+      def initialize(message, handle)
+        super(message)
+        @handle = handle
+      end
+    end
+  end
+end