+++ spec/spec_helper.rb
@@ -16,6 +16,26 @@ require_relative '../lib/deepl'
 # Lib config
 ENV['DEEPL_AUTH_KEY'] ||= 'TEST-TOKEN'
 
+# General helpers
+def build_deepl_api
+  DeepL::API.new(DeepL::Configuration.new(host: 'https://api-free.deepl.com'))
+end
+
+uri_ignoring_deepl_api_subdomain = lambda do |request1, request2|
+  deepl_api_regexp = %r{https?://api.*\.deepl\.com/}
+  uri1 = request1.uri
+  uri2 = request2.uri
+  uri1_match = uri1.match(deepl_api_regexp)
+  uri2_match = uri2.match(deepl_api_regexp)
+  if uri1_match && uri2_match
+    uri1_without_deepl_domain = uri1.gsub(uri1_match[0], '')
+    uri2_without_deepl_domain = uri2.gsub(uri2_match[0], '')
+    uri1_without_deepl_domain == uri2_without_deepl_domain
+  else
+    uri1 == uri2
+  end
+end
+
 # VCR tapes configuration
 require 'vcr'
 VCR.configure do |config|
@@ -23,12 +43,20 @@ VCR.configure do |config|
   config.hook_into :webmock
   config.filter_sensitive_data('VALID_TOKEN') { ENV.fetch('DEEPL_AUTH_KEY', nil) }
   config.default_cassette_options = {
+    # Uncomment this line when adding new tests, run the tests once, then comment it again
     # record: :new_episodes,
-    match_requests_on: %i[method uri body headers]
+    match_requests_on: [:method, uri_ignoring_deepl_api_subdomain, :body, :headers]
   }
 end
 
-# General helpers
-def build_deepl_api
-  DeepL::API.new(DeepL::Configuration.new(host: 'https://api-free.deepl.com'))
+# Test helpers
+
+def replace_env_preserving_deepl_vars
+  env_auth_key = ENV.fetch('DEEPL_AUTH_KEY', false)
+  env_server_url = ENV.fetch('DEEPL_SERVER_URL', false)
+  tmp_env = ENV.to_hash
+  ENV.clear
+  ENV['DEEPL_AUTH_KEY'] = (env_auth_key || 'VALID')
+  ENV['DEEPL_SERVER_URL'] = (env_server_url || '')
+  tmp_env
 end